requirements.txt
anthropic>=0.39.0
fpdf2>=2.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
markdown>=3.5.0
